TypeScript Translation
