+++ .gitignore
@@ -1,3 +1,4 @@
 node_modules
 test.js
 package-lock.json
+dist
