+++ package.json
@@ -4,6 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
+    "build": "rimraf dist && tsc -p tsconfig.json",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -28,6 +29,13 @@
     "express": "^4.17.1",
     "helmet": "^4.6.0",
     "uuid": "^8.3.2"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.13",
+    "@types/node": "^17.0.25",
+    "@types/uuid": "^8.3.4",
+    "rimraf": "^3.0.2",
+    "typescript": "^4.6.3"
   },
   "homepage": "https://github.com/ReAdminRBX/roblox-long-polling#readme"
 }
